+++ src/config/config.ts
@@ -1,5 +1,5 @@
 import {consts} from "./consts";
-import {ConfigDTO, KafkaConsumerConfigDTO, MySQLConfigDTO} from "../models";
+import {ConfigDTO, KafkaConsumerConfigDTO, MemcachedConfigDTO, MySQLConfigDTO} from "../models";
 
 const kafkaConsumerConfig: KafkaConsumerConfigDTO = new KafkaConsumerConfigDTO(
     consts.kafkaConsumerConfig.brokers,
@@ -16,9 +16,14 @@ const mySQLConfig: MySQLConfigDTO = new MySQLConfigDTO(
     consts.mySQLConfig.password,
     consts.mySQLConfig.database
 )
+const memCachedConfig: MemcachedConfigDTO = new MemcachedConfigDTO(
+    consts.memcachedConfig.location,
+    consts.memcachedConfig.defaultTTL
+)
 
 const config: ConfigDTO = new ConfigDTO();
 config.kafkaConsumerConfig = kafkaConsumerConfig;
 config.mySQLConfig = mySQLConfig;
+config.memcachedConfig = memCachedConfig;
 
 export default config;