+++ src/driver/index.ts
@@ -2,10 +2,14 @@ import ISQLDataSource from "./interfaces/ISQLDataSource";
 import KafkaConsumer from "./kafka/kafka";
 import MySQLDataSource from "./mysql/mysql";
 import IQueueConsumer from "./interfaces/IQueueConsumer";
+import ICache from "./interfaces/ICache";
+import Memcached from "./memcached/memcached";
 
 export {
     ISQLDataSource,
     KafkaConsumer,
     MySQLDataSource,
-    IQueueConsumer
+    IQueueConsumer,
+    ICache,
+    Memcached
 }