+++ src/driver/interfaces/ICache.ts
@@ -1,0 +1,4 @@
+export default interface ICache {
+    get(key: string): Promise<any>
+    setAsync(key:string, value: any, ttl?: number): void
+}